Move export pipeline into MainWindow.Export.cpp
