+++ Win/llvc/llvc.vcxproj
@@ -147,6 +147,7 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
     </ClCompile>
     <ClCompile Include="MainWindow.Helpers.cpp" />
+    <ClCompile Include="MainWindow.Export.cpp" />
     <ClCompile Include="$(GeneratedFilesDir)module.g.cpp" />
   </ItemGroup>
   <ItemGroup>
